skills/ecommerce-automation/requirements.txt
# 1688 选品 + 拼多多运营自动化技能 - Python 依赖包
# 安装命令：pip install -r requirements.txt

# ==================== 核心依赖 ====================
# 浏览器自动化
playwright>=1.40.0
selenium>=4.15.0

# 数据处理
pandas>=2.1.0
openpyxl>=3.1.0  # Excel 处理
xlrd>=2.0.0  # Excel 读取

# 配置文件解析
pyyaml>=6.0.1

# HTTP 请求
requests>=2.31.0
aiohttp>=3.9.0  # 异步 HTTP

# 数据库
sqlite3  # Python 内置，无需安装

# 定时任务
apscheduler>=3.10.0

# 日志增强
loguru>=0.7.0

# ==================== 可选依赖 ====================
# 图片处理（如需处理商品图片）
pillow>=10.0.0

# PDF 生成（如需生成 PDF 报表）
reportlab>=4.0.0

# 钉钉通知
dingtalk-sdk>=2.0.0

# 数据可视化（如需图表展示）
matplotlib>=3.8.0

# ==================== 开发工具 ====================
# 代码质量检查（开发时使用）
flake8>=6.1.0
black>=23.0.0

# 测试框架
pytest>=7.4.0
